requirements.txt
certifi==2021.5.30
chardet==4.0.0
easydict==1.9
idna==2.10
joblib==1.0.1
myfm==0.3.0.1
numpy~=1.19.5
pandas==1.3.0
py4j==0.10.9
pybind11==2.6.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
threadpoolctl==2.1.0
torch==1.9.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
matplotlib~=3.4.2
hickle==4.0.4